move BSGS-k unit test closer to implementation
